fix: Do not self import @calcom/ui
