+++ packages/types/CredentialOwner.d.ts
@@ -1,0 +1,7 @@
+export type CredentialOwner = {
+  name: string | null;
+  avatar?: string | null;
+  teamId?: number;
+  credentialId?: number;
+  readOnly?: boolean;
+};
