+++ packages/ui/components/apps/AppCard.tsx
@@ -13,11 +13,12 @@ import { useLocale } from "@calcom/lib/hooks/useLocale";
 import type { UserAdminTeams } from "@calcom/lib/server/repository/user";
 import type { AppFrontendPayload as App } from "@calcom/types/App";
 import type { CredentialFrontendPayload as Credential } from "@calcom/types/Credential";
-import type { ButtonProps } from "@calcom/ui";
-import { Badge, showToast } from "@calcom/ui";
 import classNames from "@calcom/ui/classNames";
 
+import { Badge } from "../badge/Badge";
+import type { ButtonProps } from "../button";
 import { Button } from "../button";
+import { showToast } from "../toast/showToast";
 
 interface AppCardProps {
   app: App;
