+++ packages/ui/components/avatar/UserAvatar.tsx
@@ -2,8 +2,9 @@ import { getPlaceholderAvatar } from "@calcom/lib/defaultAvatarImage";
 import { getUserAvatarUrl } from "@calcom/lib/getAvatarUrl";
 import type { User } from "@calcom/prisma/client";
 import type { UserProfile } from "@calcom/types/UserProfile";
-import { Avatar } from "@calcom/ui";
 import classNames from "@calcom/ui/classNames";
+
+import { Avatar } from "./Avatar";
 
 type Organization = {
   id: number;
