+++ packages/ui/components/disconnect-calendar-integration/DisconnectIntegration.tsx
@@ -1,6 +1,9 @@
 import { useLocale } from "@calcom/lib/hooks/useLocale";
-import type { ButtonProps } from "@calcom/ui";
-import { Button, ConfirmationDialogContent, Dialog, DialogTrigger } from "@calcom/ui";
+
+import type { ButtonProps } from "../button/Button";
+import { Button } from "../button/Button";
+import { ConfirmationDialogContent } from "../dialog/ConfirmationDialogContent";
+import { Dialog, DialogTrigger } from "../dialog/Dialog";
 
 export const DisconnectIntegrationComponent = ({
   label,
