+++ packages/ui/components/form/toggleGroup/ToggleGroup.tsx
@@ -1,8 +1,9 @@
 import * as RadixToggleGroup from "@radix-ui/react-toggle-group";
 import type { ReactNode } from "react";
 
-import { Tooltip } from "@calcom/ui";
 import classNames from "@calcom/ui/classNames";
+
+import { Tooltip } from "../../tooltip/Tooltip";
 
 interface ToggleGroupProps extends Omit<RadixToggleGroup.ToggleGroupSingleProps, "type"> {
   options: {
